New Levels and Tutorials
Updated the tutorial level so that it's an actual tutorial (some text/images/etc). Also added 2 more tutorials to go over special platforms (selective and bouncy). Additional levels added outside of the tutorials. Now a total of 9 levels.

Also added a "Next Level" option to the end-level popup so that you don't have to go to the chapter select every time you win. Will only show if you get 3 stars on the level.

Also added a "Quit to Desktop" button on the main menu because that's a thing that games have.