Menu Updates
Gave the main menu and level selection menu a much needed UI update. Now has a background image for the main menu, highlighting when hovering over the different options, and just generally polished it up a bit. Set up the ability to change the default cursor and have a separate one for when hovering over the menu options. Not using it now, but might make custom cursors sometime in the future.