More Menu Updates and Star Sprites!
Sprites for stars have been added so now there is a visual indicator of the stars you earned instead of seeing the text "Stars:3". This was updated in both the Level Select Menu and the Level End Menu.

The different menus (Main, Level Select, Pause, and Level End) have all been separated into their own classes. This doesn't change anything in the game, but is more cleanly coded and isolates the different menus from eachother (which is how I should have done it in the first place, but oh well).