Star Bug Fix, Level Lock Icon, & General Code Cleaning
Fixed a bug in the end-level score menu where the stars were not updating properly. Earned stars from the previous levels were carrying over to the next. The stars now show exactly to what you earned for that specific level.

Added a lock icon to the Level Select Menu for locked levels. This disappears once you unlock the level.

Cleaned up the code a bit for the Drone movement. Switched from using FixedUpdate to using the delta time. Makes it a bit smoother overall and I don't have to use such small values for the movement speed so it'll be easier for me to adjust it as needed.

Added a bunch of comments for the drone controller and a few other classes so that I actually know what I'm looking in the future.