Changed Drone Movement
The drone's movement is now based off of the rigidbody velocity rather than just updating the position of the gameobject. This was done to resolve some issues with collision when moving diagonally (you could go through some objects at the corners before). Will need to tweak and play around with it as the movement feels a bit heavy, but for now it's a good start.

Also added an fps counter during gamplay so that i could test how it runs on different computers and see if anything causes issues with stability.

A quick restart was also added. Press 'R' to immediately restart the level.

New test scene. I am working on implementing a new rope system based on Verlet integration. This should eventually make the rope more realistic and feel more loose and swingy. Right now, the rope is using a spring joint which works kind of well, but basically disregards the gravity since it's always trying to pull you towards the drone. It doesn't always do what I expect it to do. So I'm trying to move away from that and use a more realistic rope approach.